Separate the loss as a function
